+++ CHANGELOG.md
@@ -4,7 +4,8 @@ CHANGELOG
 7.4
 ---
 
-* Add `MermaidDumper` to dump Role Hierarchy graphs in the Mermaid.js flowchart format
+ * Add `MermaidDumper` to dump Role Hierarchy graphs in the Mermaid.js flowchart format
+ * Deprecate `PersistentTokenInterface::getClass()`, the user class will be removed from the remember-me cookie in 8.0
 
 7.3
 ---